RvFeaturesSettings: Add kernel compatibility check
Introduced a method to check the kernel version using the `uname -r` command to ensure that the features displayed in the UI are compatible with the running kernel. If the kernel version does not match the required version, the RvFeatures category is removed from the settings.

This prevents the user from accessing incompatible features based on the kernel version.

Signed-off-by: Rve27 <rve27github@gmail.com>